removed management commands from daemon processes, updated systemstat daemon,
